fixed the sidebar toggle not adjusting cards issue
