+++ client/src/constants/theme.ts
@@ -2,7 +2,7 @@ import { createTheme } from '@mui/material';
 
 export const borderRadius = 10;
 export const borderWidth = 3;
-export const leftContentPadding = 138;
+export const leftContentPadding = 64;
 export const rightContentPadding = 64;
 export const contentPadding = 15;
 export const inventoryDropzoneOpacity = 0.1;
